30 [ADD] save decision tree
DONE:
    - Add a function to save the tree in a file
    - Add a function to import the tree from a file
    - Modify the test to works with the save format

TODO:
    - Merge dev

issue number : #30

+++ DecisionTree/Node.py
@@ -19,7 +19,7 @@ class Node:
             column_index (int) : Index of the decision column in the original data.
             operator (str) : '=' if the decision is made on a binary value else '<='
 
-            value (int) : The most represented class by this node.
+            value (float) : The most represented class by this node.
         """
